fix form look style
