+++ web-app/src/components/settings/adminPanelContent/CategoriesTable.tsx
@@ -1,8 +1,8 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import CategoryList from "./Category";
 import { useQuery } from "@apollo/client";
-import CategoryEditionForm from "./CategoryEditionForm";
-import CreateCategoriesForm from "@/components/forms/CreateCategoriesForm";
+import EditCategoryForm from "../../forms/edit/EditCategoryForm";
+import CreateCategoryForm from "@/components/forms/create/CreateCategoryForm";
 import { Category, GetCategoriesQuery } from "@/gql/generate/graphql";
 import { GET_CATEGORIES } from "@/gql/requests/queries";
 
@@ -11,6 +11,15 @@ function CategoriesTable() {
   const [isCreationPanelAdmin, setIsCreationPanelAdmin] = useState(false);
   const [currentCategory, setCurrentCategory] = useState<Category | null>(null);
   const [searchCategory, setSearchCategory] = useState("");
+
+  useEffect(() => {
+    if (isEditionPanelAdmin || isCreationPanelAdmin) {
+      document.body.classList.add("overflow-hidden");
+    } else {
+      document.body.classList.remove("overflow-hidden");
+    }
+    return () => document.body.classList.remove("overflow-hidden");
+  }, [isEditionPanelAdmin, isCreationPanelAdmin]);
 
   const openEditionPanelAdmin = (category: Category) => {
     setCurrentCategory(category);
@@ -95,9 +104,9 @@ function CategoriesTable() {
           </tbody>
         </table>
         {isCreationPanelAdmin && (
-          <div className="w-screen h-screen bg-white fixed inset-0 px-2 z-10 overflow-hidden flex items-center justify-center">
-            <div className="pb-12 border border-gray-300 rounded-2xl">
-              <CreateCategoriesForm
+          <div className="fixed inset-0 z-50 bg-gray-800 backdrop-blur-sm bg-opacity-50 flex items-center justify-center">
+            <div className="bg-white p-2 rounded-lg shadow-lg overflow-y-auto">
+              <CreateCategoryForm
                 setIsCreationPanelAdmin={setIsCreationPanelAdmin}
                 refetchCategoryData={refetch}
               />
@@ -105,13 +114,15 @@ function CategoriesTable() {
           </div>
         )}
         {isEditionPanelAdmin && currentCategory && (
-          <div className="w-screen h-screen fixed inset-0 px-2 z-10 overflow-hidden flex items-center justify-center">
-            <CategoryEditionForm
-              key={currentCategory.id}
-              category={currentCategory}
-              setIsEditionPanelAdmin={closeEditionPanelAdmin}
-              refetch={refetch}
-            />
+          <div className="fixed inset-0 z-50 bg-gray-800 backdrop-blur-sm bg-opacity-50 flex items-center justify-center">
+            <div className="bg-white p-2 rounded-lg shadow-lg overflow-y-auto">
+              <EditCategoryForm
+                key={currentCategory.id}
+                category={currentCategory}
+                setIsEditionPanelAdmin={closeEditionPanelAdmin}
+                refetch={refetch}
+              />
+            </div>
           </div>
         )}
       </div>
